requirements.txt
streamlit
pandas
numpy
plotly
prophet
textblob
requests
joblib
scikit-learn
pulp
yagmail
python-dotenv
